No need for threadsafe
git-svn-id: svn+ssh://solsvn.oma.be/home/sidc/ssr/SWHV_repos/trunk/client/jhv-3d-wcs@5959 b4e469a2-07ce-4b26-9273-4d7d95a670c7

+++ src/jhv/src/org/helioviewer/jhv/viewmodel/view/jp2view/JP2View.java
@@ -270,7 +270,7 @@ public class JP2View extends AbstractView {
         long delta = currentTime - frameCountStart;
 
         if (delta > 1000) {
-            frameRate = 1000.f * frameCount / delta;
+            frameRate = 1000 * frameCount / (float) delta;
             frameCount = 0;
             frameCountStart = currentTime;
         }
